+++ Telepath/TPTrackerKeyboard.h
@@ -10,6 +10,6 @@
 
 @interface TPTrackerKeyboard : NSObject
 @property (nonatomic, strong) NSArray *modifierKeys;
-@property (readonly) uint currentEvents;
-@property (readonly) uint totalEvents;
+@property (readonly) NSInteger currentEvents;
+@property (readonly) NSInteger totalEvents;
 @end
